Add .editorconfig, ignore TestResults, default non-verbose, fix bugs
Line endings are LF by default now.

ToStringExt is no longer generic, it doesn't need to be generic and
 actually caused a bug because it was generic where it didn't recognize
 the actual type of multiple objects, because they were passed as
 System.Object and the type recognition code was using typeof(T).
Now it uses .GetType() so it gets the real object type.
Updated the doc comment to reflect this.
Fixed a bug where ToStringExt did not handle null.

The new test project is in the next commit, but ignoring the
 TestResults directory here.
Ignore the new test project in the main project.

Verbose is now off by default.

Fixed the Root Namespace to be correct.